+++ .changeset/large-chefs-boil.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-scaffolder-react': minor
+---
+
+move useEventStream, TaskSteps, TaskBorder and TaskLogStream into scaffolder-react
