+++ plugins/scaffolder-react/package.json
@@ -58,6 +58,7 @@
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "4.0.0-alpha.57",
+    "@react-hookz/web": "^20.0.0",
     "@rjsf/core": "^3.2.1",
     "@rjsf/core-v5": "npm:@rjsf/core@5.1.0",
     "@rjsf/material-ui": "^3.2.1",
@@ -66,10 +67,12 @@
     "@rjsf/validator-ajv8": "5.1.0",
     "@types/json-schema": "^7.0.9",
     "classnames": "^2.2.6",
+    "humanize-duration": "^3.25.1",
     "immer": "^9.0.1",
     "json-schema": "^0.4.0",
     "json-schema-library": "^7.3.9",
     "lodash": "^4.17.21",
+    "luxon": "^3.0.0",
     "qs": "^6.9.4",
     "react-use": "^17.2.4",
     "use-immer": "^0.8.0",
@@ -91,7 +94,9 @@
     "@testing-library/jest-dom": "^5.10.1",
     "@testing-library/react": "^12.1.3",
     "@testing-library/react-hooks": "^8.0.0",
-    "@testing-library/user-event": "^14.0.0"
+    "@testing-library/user-event": "^14.0.0",
+    "@types/humanize-duration": "^3.18.1",
+    "@types/luxon": "^3.0.0"
   },
   "files": [
     "dist"
