+++ plugins/scaffolder-react/src/components/TaskLogStream/index.ts
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { TaskSteps } from './TaskSteps';
+export { TaskLogStream } from './TaskLogStream';
