+++ plugins/scaffolder-react/src/components/TaskSteps/StepIcon.tsx
