+++ plugins/scaffolder-react/src/components/TaskSteps/TaskSteps.tsx
@@ -27,11 +27,20 @@ import { type Step } from '@backstage/plugin-scaffolder-react';
 import { StepIcon } from './StepIcon';
 import { StepTime } from './StepTime';
 
-interface StepperProps {
+/**
+ *
+ * @public
+ */
+export interface StepperProps {
   steps: (TaskStep & Step)[];
   activeStep?: number;
 }
 
+/**
+ * The visual stepper of the task event stream
+ *
+ * @public
+ */
 export const TaskSteps = (props: StepperProps) => {
   return (
     <MuiStepper
