+++ plugins/scaffolder-react/src/components/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,11 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export * from './extensions';
-export * from './types';
-export * from './secrets';
-export * from './api';
-export * from './hooks';
-export * from './layouts';
-export * from './components';
+export { TaskSteps, type StepperProps } from './TaskSteps';
+export { TaskBorder } from './TaskBorder';
+export { TaskLogStream } from './TaskLogStream';
