+++ plugins/scaffolder/src/next/OngoingTask/OngoingTask.tsx
@@ -17,17 +17,17 @@ import React, { useEffect, useMemo, useState, useCallback } from 'react';
 import { Page, Header, Content, ErrorPanel } from '@backstage/core-components';
 import { useNavigate, useParams } from 'react-router-dom';
 import { Box, makeStyles, Paper } from '@material-ui/core';
-import { TaskSteps } from './TaskSteps';
-import { TaskBorder } from './TaskBorder';
-import { TaskLogStream } from './TaskLogStream';
+import {
+  TaskSteps,
+  TaskBorder,
+  TaskLogStream,
+  ScaffolderTaskOutput,
+  useTaskEventStream,
+} from '@backstage/plugin-scaffolder-react';
 import { nextSelectedTemplateRouteRef } from '../routes';
 import { useRouteRef } from '@backstage/core-plugin-api';
 import qs from 'qs';
 import { ContextMenu } from './ContextMenu';
-import {
-  ScaffolderTaskOutput,
-  useTaskEventStream,
-} from '@backstage/plugin-scaffolder-react';
 import { DefaultTemplateOutputs } from '@backstage/plugin-scaffolder-react/alpha';
 
 const useStyles = makeStyles({
